+++ HW4/Spreadsheet_Isaac_Dahle/Commands/CommandManager.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Cass Dahle 11775278.
 // Licensed under the GPL v3.0 License. See LICENSE in the project root for license information.
 
+using System.Collections.Generic;
+
 namespace SpreadsheetEngine;
 
 /// <summary>
